pyproject.toml
[tool.poetry]
name = "mag-lab-data-toolkit"
version = "0.1.0"
description = ""
authors = ["mhairifin <m.finlayson@imperial.ac.uk>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pydantic = "^2.6.1"
pytest = "^8.3.1"
pyyaml = "^6.0.1"
cdflib = "^1.3.1"
xarray = "^2024.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
